+++ src/ServiceProvider.php
@@ -22,7 +22,7 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
 {
     use ServiceDocBuilder;
 
-    public function register()
+    public function boot()
     {
         // Add our database extensions.
         $this->app->resolving('db.schema', function (DbSchemaExtensions $db){
@@ -90,5 +90,8 @@ class ServiceProvider extends \Illuminate\Support\ServiceProvider
         $this->app->resolving('df.system.table_model_map', function (SystemTableModelMapper $df) {
             $df->addMapping('sql_db_config', SqlDbConfig::class);
         });
+
+        // add migrations
+        $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
     }
 }