+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+
 ### Changed
+- Batch requests now report errors consistently using BatchException
+
 ### Fixed
 
 ## [0.7.1] - 2017-01-25